fix metrics codeql circular import
